Double precision overlay

+++ Rendering/shaders/overlay_vertex.glsl
@@ -1,14 +1,15 @@
 R"(
 #version 460 core
+#extension GL_EXT_vertex_attrib_64bit : require
 
-layout (location = 0) in vec2 position;
+layout (location = 0) in dvec2 position;
 
 uniform dvec2 viewportCenter;
 uniform double viewportBreadth;
 
 out float index;
 
-vec2 viewportToGL(vec2 p) {
+vec2 viewportToGL(dvec2 p) {
     return vec2((p - viewportCenter) / viewportBreadth); //Cast to float vec
 }
